% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{generate_matern_basis2}
\alias{generate_matern_basis2}
\title{generate_matern_basis2}
\usage{
generate_matern_basis2(
  grids,
  region_idx_list,
  L_vec,
  scale = 2,
  nu = 1/5,
  show_progress = FALSE
)
}
\description{
generate_matern_basis2
}
