% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Low_to_high}
\alias{Low_to_high}
\title{Map basis coefficients in matrix form to the image space}
\usage{
Low_to_high(Low_mat, p, Phi_Q, region_idx, L_idx)
}
\description{
Map basis coefficients in matrix form to the image space
}
