% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Low_to_high_vec}
\alias{Low_to_high_vec}
\title{Map basis coefficients to the image space}
\usage{
Low_to_high_vec(Low_vec, p, Phi_Q, region_idx, L_idx)
}
\description{
Map basis coefficients to the image space
}
