% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{plot_multi_img}
\alias{plot_multi_img}
\title{plot_multi_img}
\usage{
plot_multi_img(
  list_of_image,
  grids_df,
  n_img_per_row = 3,
  col_bar = NULL,
  layout_matrix = NULL,
  labels = NULL,
  font_size = 20,
  color_mode = "BWR",
  legend_position = "right"
)
}
\description{
plot_multi_img
}
