% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{Unconfounded_mediation_data_constrained_eta}
\alias{Unconfounded_mediation_data_constrained_eta}
\title{generate image mediation data with unobserved confounders}
\usage{
Unconfounded_mediation_data_constrained_eta(
  beta,
  alpha,
  n,
  Q,
  D,
  lambda,
  nu = NULL,
  lambda_nu = 0,
  q = 2,
  sigma_M = 0.1,
  sigma_Y = 0.1,
  nu_scale = 1,
  sigma_eta = 0.1,
  gamma = 2
)
}
\description{
generate image mediation data with unobserved confounders
}
